requirements.txt
# DualTreeVLA dependencies
# Install order matters: torch first, then flash-attn

# ── Core ──────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── LLM / Tokenizer ───────────────────────────────────────────────
transformers>=4.40.0,<4.45.0
accelerate>=0.28.0
timm>=0.9.0          # InternVL3 ViT 依赖（modeling_intern_vit.py 中使用 timm.layers.DropPath）

# ── DeepSpeed (multi-GPU) ─────────────────────────────────────────
deepspeed>=0.14.0,<0.15.0

# ── Flash Attention (optional but recommended on A6000/Ampere+)
# Install separately after torch:
#   pip install flash-attn --no-build-isolation
# flash-attn>=2.5.0

# ── Data / IO ─────────────────────────────────────────────────────
h5py>=3.9.0
opencv-python>=4.8.0
pyyaml>=6.0
einops>=0.7.0

# ── Weights & Biases (optional, for visualisation) ──────────────────────
wandb>=0.17.0

# ── Build tools（flash-attn / deepspeed 编译依赖，无 sudo 时替代 apt-get）──
ninja>=1.11.0
tqdm>=4.66.0
cmake>=3.26.0

# ── Eval server/client (WebSocket) ────────────────────────────────
websockets>=12.0

# ── Utilities ─────────────────────────────────────────────────────
numpy>=1.24.0,<2.0
tqdm>=4.65.0
